+++ app/controllers/pages.js
@@ -4,13 +4,20 @@ var ND_Page = require('../models/ND_Page');
 var pages = {
     list: function (req, res) {
         var parent_id = req.query.parent_id,
-            searchCriteria = parent_id ? {parent_id: parent_id} : {};
+            searchCriteria = parent_id ? {parent_id: parent_id} : {},
+            response = {};
 
         ND_Page.find(searchCriteria, function (err, nd_pages) {
             if (err) {
-                console.log(err);
+                response['status'] = "error";
+                response['content'] = err;
+
+                res.json(response);
             } else {
-                res.json(nd_pages);
+                response['status'] = "ok";
+                response['content'] = nd_pages;
+
+                res.json(response);
             }
         });
     },
@@ -20,7 +27,8 @@ var pages = {
             is_deleted = req.query.is_deleted || 1,
             is_visible = req.query.is_visible || 1,
             name = req.query.name,
-            created = new Date();
+            created = new Date(),
+            response = {};
 
 
         var nd_page = new ND_Page({
@@ -33,10 +41,14 @@ var pages = {
 
         nd_page.save(function (err) {
             if (err) {
-                console.log(err);
-                res.json(err);
+                response['status'] = "error";
+                response['content'] = err;
+
+                res.json(response);
             } else {
-                res.json('ok');
+                response['status'] = "ok";
+
+                res.json(response);
             }
         });
     },
@@ -44,13 +56,20 @@ var pages = {
 
     read: function (req, res) {
         var parent_id = req.query.parent_id,
-            searchCriteria = parent_id ? {parent_id: parent_id} : {};
+            searchCriteria = parent_id ? {parent_id: parent_id} : {},
+            response = {};
 
         ND_Page.find(searchCriteria, function (err, nd_pages) {
             if (err) {
-                console.log(err);
+                response['status'] = "error";
+                response['content'] = err;
+
+                res.json(response);
             } else {
-                res.json(nd_pages);
+                response['status'] = "ok";
+                response['content'] = nd_pages;
+
+                res.json(response);
             }
         });
     },
@@ -62,12 +81,18 @@ var pages = {
 
 
     remove: function (req, res) {
+        var response = {};
+
         ND_Page.remove({_id: req.query._id}, function (err) {
             if (err) {
-                res.json(err);
-            }
-            else {
-                res.json("OK");
+                response['status'] = "error";
+                response['content'] = err;
+
+                res.json(response);
+            } else {
+                response['status'] = "ok";
+
+                res.json(response);
             }
         });
     }